Uploaded the java files and class files.
The source code (.java file) includes java's awt and swing packages to create a window allowing users to enter the key (passphrase) and select the image to be encrypted.
The image is first converted into an array of bytes which is further XORed with the unique passkey and stored back into the byte array. The principle of XOR is used to encrypt and decrypt the image.